# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SN Auto Parts API
  description: |
    Feature-config driven e-commerce API for auto parts.
    
    ## Authentication
    All authenticated endpoints require a session token passed via:
    - Cookie: `auth-token=<token>`
    - Header: `Authorization: Bearer <token>`
    
    ## Authorization
    APIs are generic - authorization is determined by the authenticated user's role 
    and feature configuration. The same endpoint may return different data or 
    permissions based on the user's role.
    
    ## Feature-Based Access
    Each endpoint requires specific features to be enabled in the user's role configuration.
    If a user lacks the required feature, they receive a 403 Forbidden response.
  version: 1.0.0
  contact:
    name: SN Auto Parts
    email: support@snautoparts.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.snautoparts.com/api/v1
    description: Production server
...
paths:
...
  /checkout/payment-intent:
    post:
      tags: [Checkout]
      summary: Create Stripe payment intent
      operationId: createPaymentIntent
      security:
        - sessionAuth: []
      x-feature-required: cart.checkout
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                idempotencyKey:
                  type: string
      responses:
        '200':
          description: Payment intent created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/PaymentIntent'
        '400':
          $ref: '#/components/responses/BadRequest'
  /checkout/orders:
    post:
      tags: [Checkout]
      summary: Create order
      operationId: createOrder
      security:
        - sessionAuth: []
      x-feature-required: cart.checkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataWrapper'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  # ============================================
  # ORDERS ENDPOINTS
  # ============================================
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      description: |
        Returns orders based on user's role:
        - CUSTOMER: Only own orders (requires orders.viewOwn)
        - MANAGER/ADMIN: All orders (requires orders.viewAll)
      operationId: listOrders
      security:
        - sessionAuth: []
      x-feature-required: orders.viewOwn|orders.viewAll
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/OrderSummary'
...
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session token (also accepted as auth-token cookie)
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    DataWrapper:
      type: object
      properties:
        data: {}
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
        details:
          type: object

    # Auth Schemas
...
    PaymentIntent:
      type: object
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
        amount:
          type: number
        subtotal:
          type: number
        shippingAmount:
          type: number
        taxAmount:
          type: number
    CreateOrderRequest:
      type: object
      required: [shippingMethod]
      properties:
        addressId:
          type: string
        shippingAddress:
          $ref: '#/components/schemas/CreateAddressRequest'
        shippingMethod:
          type: string
        paymentIntentId:
          type: string
        idempotencyKey:
          type: string
        notes:
          type: string

    # Order Schemas
    OrderStatus:
      type: string
      enum: [PENDING, CONFIRMED, PROCESSING, SHIPPED, DELIVERED, CANCELLED, REFUNDED]
    OrderSummary:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        paymentStatus:
          type: string
        totalAmount:
          type: string
        createdAt:
          type: string
          format: date-time
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            email:
              type: string
        _count:
          type: object
          properties:
            items:
              type: integer
    Order:
      allOf:
        - $ref: '#/components/schemas/OrderSummary'
        - type: object
          properties:
            subtotal:
              type: string
            shippingAmount:
              type: string
            taxAmount:
              type: string
            shippingMethod:
              type: string
            shippingCarrier:
              type: string
            trackingNumber:
              type: string
            notes:
              type: string
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItem'
            timeline:
              type: array
              items:
                $ref: '#/components/schemas/OrderTimelineEntry'
    OrderItem:
      type: object
      properties:
        id:
          type: string
        sku:
          type: string
        name:
          type: string
        price:
          type: string
        quantity:
          type: integer
        totalPrice:
          type: string
        product:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            imageUrl:
              type: string
            slug:
              type: string
    OrderTimelineEntry:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
...
    CreateAddressRequest:
      type: object
      required: [firstName, lastName, street, city, state, zipCode]
      properties:
        label:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        street:
          type: string
        apartment:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
          default: US
        phone:
          type: string
        isDefault:
          type: boolean
...